Added finally trpc and fixed eslint errors
