+++ package.json
@@ -18,19 +18,17 @@
     "@trpc/client": "^10.35.0",
     "@trpc/next": "^10.35.0",
     "@trpc/server": "^10.35.0",
-    "superjson": "^1.13.1"
-  },
-  "devDependencies": {
-    "@types/eslint": "^8.21.1",
+    "superjson": "^1.13.1",
+    "@types/eslint": "^8.37.0",
     "@types/node": "^18.14.0",
     "@types/prettier": "^2.7.2",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.53.0",
-    "@typescript-eslint/parser": "^5.53.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "autoprefixer": "^10.4.7",
-    "eslint": "^8.34.0",
-    "eslint-config-next": "^13.2.1",
+    "eslint": "^8.40.0",
+    "eslint-config-next": "^13.4.2",
     "postcss": "^8.4.14",
     "prettier": "^2.8.1",
     "prettier-plugin-tailwindcss": "^0.2.1",
