+++ src/pages/_app.tsx
@@ -1,9 +1,10 @@
-import { type AppType } from "next/dist/shared/lib/utils";
+import { type AppType } from "next/app";
 
 import "~/styles/globals.css";
+import { api } from "~/lib/utils/api";
 
 const MyApp: AppType = ({ Component, pageProps }) => {
   return <Component {...pageProps} />;
 };
 
-export default MyApp;
+export default api.withTRPC(MyApp);
